Revert "[BugFix][v0.22.1] Handle wrapped reasoning parsers in usage accounting (#10790)"
This reverts commit c6b62d0e9a9b6a2f1dc61728fe93cc0bd91f5fde.
