+++ vllm_ascend/patch/platform/__init__.py
@@ -29,7 +29,6 @@ if not is_310p():
 else:
     import vllm_ascend.patch.platform.patch_mamba_config_310  # noqa
 import vllm_ascend.patch.platform.patch_minimax_m2_config  # noqa
-import vllm_ascend.patch.platform.patch_chat_usage_accounting  # noqa
 import vllm_ascend.patch.platform.patch_minimax_usage_accounting  # noqa
 import vllm_ascend.patch.platform.patch_glm_tool_call_streaming  # noqa
 import vllm_ascend.patch.platform.patch_glm47_tool_call_parser  # noqa
